# server/mcp/openapi.json
{
  "components": {
    "parameters": {
...
      "costLimit": {
        "description": "Cost limit in configured currency (default EUR) or CO2 limit in g/kWh",
        "example": 0.25,
        "in": "path",
        "name": "cost",
        "required": true,
        "schema": {
          "type": "number"
        }
      },
...
      "enable": {
        "description": "Charging mode.",
        "in": "path",
        "name": "enable",
        "required": true,
        "schema": {
          "enum": [
            "true",
            "false",
            "1",
            "0"
          ],
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "BooleanResult": {
        "content": {
          "application/json": {
            "schema": {
              "type": "boolean"
            }
          }
        },
        "description": "Success - Boolean result"
      },
...
      "Unauthorized": {
        "content": {
          "text/plain": {
            "schema": {
              "enum": [
                "Unauthorized"
              ],
              "type": "string"
            }
          }
        },
        "description": "Unauthorized - Login and try again"
      }
    },
    "schemas": {
...
      "ChangePassword": {
        "properties": {
          "current": {
            "$ref": "#/components/schemas/Password"
          },
          "new": {
            "$ref": "#/components/schemas/Password"
          }
        },
        "type": "object"
      },
...
      "Password": {
        "description": "Admin password",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Long-lived API key (prefixed `evcc_`, managed via /auth/apikey).\nGrants access to /api/db/* without requiring the X-Admin-Password header.\n",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "auth",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://github.com/evcc-io/evcc"
    },
    "description": "Solar charging. Super simple.",
    "title": "evcc",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/apikey": {
      "get": {
        "description": "Reports whether an API key has been generated. The key itself is\nonly returned once at creation time (POST), never on subsequent\nreads.\n",
        "operationId": "getApiKeyStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "configured": {
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "API key status",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "description": "Generates a fresh API key, replacing any existing one. The\nreturned key is shown only once; store it immediately.\n\nEven when the request is authenticated via API key (Bearer), the\nadmin password must be supplied in the request body to prevent a\nleaked key from rotating itself. The password check is skipped\nwhen the server is started with `--disable-auth`.\n",
        "operationId": "regenerateApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "$ref": "#/components/schemas/Password"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "key": {
                      "description": "The new API key (cleartext, shown once).",
                      "example": "evcc_aB3xYz7k0Pq2sN5mWtR4",
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Invalid admin password"
          },
          "403": {
            "description": "Forbidden in demo mode"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Generate or rotate the API key",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Administrator login. Returns authorization cookie required for all protected endpoints.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "$ref": "#/components/schemas/Password"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Set-Cookie": {
                "description": "The authorization cookie",
                "schema": {
                  "example": "auth=ey...qY; Path=/; Expires=Fri, 04 Apr 2025 14:35:37 GMT; HttpOnly; SameSite=Strict",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid password provided"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/password": {
      "put": {
        "description": "Changes the admin password.",
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid password provided"
          }
        },
        "summary": "Change admin password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/batterydischargecontrol/{enable}": {
      "post": {
        "description": "Prevent home battery discharge during vehicle fast charging.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "operationId": "setBatteryDischargeControl",
        "parameters": [
          {
            "$ref": "#/components/parameters/enable"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BooleanResult"
          }
        },
        "summary": "Control battery-discharge",
        "tags": [
          "battery"
        ]
      }
    },
...
    "/batterygridchargelimit/{cost}": {
      "post": {
        "description": "Charge home battery from grid when price or emissions are below the threshold. Uses price if a dynamic tariff exists. Uses emissions if a CO₂-tariff is configured. Ignored otherwise.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "operationId": "setBatteryGridChargeLimit",
        "parameters": [
          {
            "$ref": "#/components/parameters/costLimit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        },
        "summary": "Set battery grid charge limit",
        "tags": [
          "battery"
        ]
      }
    },
...
    "/buffersoc/{soc}": {
      "post": {
        "description": "Set battery buffer SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "operationId": "setBufferSoc",
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        },
        "summary": "Set battery buffer SoC",
        "tags": [
          "battery"
        ]
      }
    },
    "/bufferstartsoc/{soc}": {
      "post": {
        "description": "Set battery buffer start SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "operationId": "setBufferStartSoc",
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        },
        "summary": "Set battery buffer start SoC",
        "tags": [
          "battery"
        ]
      }
    },
    "/db/backup": {
      "get": {
        "description": "Downloads the SQLite database as a backup file. Session users must supply the admin password in the X-Admin-Password header. API key holders via Bearer token are exempt.",
        "operationId": "downloadBackup",
        "parameters": [
          {
            "$ref": "#/components/parameters/adminPassword"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "SQLite database file"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Download database backup",
        "tags": [
          "db"
        ]
      }
    },
    "/db/reset": {
      "post": {
        "description": "Selectively deletes sessions and/or settings from the database. Session users must supply the admin password in the X-Admin-Password header. API key holders via Bearer token are exempt. The instance restarts after a successful reset.",
        "operationId": "resetDatabase",
        "parameters": [
          {
            "$ref": "#/components/parameters/adminPassword"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "sessions": {
                    "description": "Delete all charging sessions",
                    "type": "boolean"
                  },
                  "settings": {
                    "description": "Delete all settings, configs, and meters",
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Reset successful, instance is restarting"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reset database",
        "tags": [
          "db"
        ]
      }
    },
    "/db/restore": {
      "post": {
        "description": "Restores the database from a previously downloaded backup file. Session users must supply the admin password in the X-Admin-Password header. API key holders via Bearer token are exempt. The instance restarts after a successful restore.",
        "operationId": "restoreBackup",
        "parameters": [
          {
            "$ref": "#/components/parameters/adminPassword"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "description": "SQLite database backup file",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Restore successful, instance is restarting"
          },
          "400": {
            "description": "Invalid file or missing fields"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Restore database backup",
        "tags": [
          "db"
        ]
      }
    },
    "/gridsessions": {
      "get": {
        "description": "Returns a list of HEMS grid limitation events.",
        "operationId": "getGridSessions",
        "parameters": [
          {
            "description": "Response format (default json)",
            "in": "query",
            "name": "format",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          },
          {
            "description": "Language (defaults to accept header)",
            "in": "query",
            "name": "lang",
            "schema": {
              "example": "de",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GridSessions"
                }
              },
              "text/csv": {
                "schema": {
                  "description": "Download csv-file",
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "Grid limitation sessions",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/prioritysoc/{soc}": {
      "post": {
        "description": "Set battery priority SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "operationId": "setPrioritySoc",
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        },
        "summary": "Set battery priority SoC",
        "tags": [
          "battery"
        ]
      }
    },
    "/residualpower/{power}": {
      "post": {
        "description": "Set grid connection operating point.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/reference/configuration/site#residualpower"
        },
        "operationId": "setResidualPower",
        "parameters": [
          {
            "$ref": "#/components/parameters/power"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        },
        "summary": "Set grid residual power",
        "tags": [
          "battery"
        ]
      }
    },
...
  },
...
}